refactor: use centralized file browser exports
