+++ src/components/fileBrowser/index.ts
@@ -1,0 +1,2 @@
+export { default as FileListView } from "./ListView"
+export { default as FileGridView } from "./GridView"
